pyproject.toml
[tool.poetry]
name = "sbc-valuation"
version = "0.1.0"
description = ""
authors = ["Ahet Besiroglu <ahmetybesiroglu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
beautifulsoup4 = "^4.12.3"
certifi = "^2024.8.30"
charset-normalizer = "^3.3.2"
frozendict = "^2.4.4"
html5lib = "^1.1"
idna = "^3.8"
inflection = "^0.5.1"
lxml = "^5.3.0"
multitasking = "^0.0.11"
numpy = "^2.1.1"
pandas = "^2.2.2"
peewee = "^3.17.6"
platformdirs = "^4.3.2"
pyairtable = "^2.3.3"
pydantic = "^2.9.1"
pydantic-core = "^2.23.3"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
requests = "^2.32.3"
scipy = "^1.14.1"
six = "^1.16.0"
soupsieve = "^2.6"
typing-extensions = "^4.12.2"
tzdata = "^2024.1"
urllib3 = "^2.2.2"
webencodings = "^0.5.1"
yfinance = "^0.2.43"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
black = "^24.8.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
